MERGE

See merge request s08-webmobile2-sub2/S08P12C109!363

+++ front/src/components/TitleCompetition/item/NestedCommentListItem.vue
@@ -2,12 +2,12 @@
   <div>
     <li>
       <div class="flex items-center mb-2">
-        <div class="w-3 h-3 rounded-full mr-2">
+        <div class="w-3 h-3 rounded-full mr-2" @click="goToProfile">
           <img :src="require(`@/assets/${profile_image}`)" alt="프로필 이미지" class="rounded-full" />
         </div>
         <p class="text-xs mr-2 font-bold">{{ nickname }}</p>
-        <p class="text-xs mr-1">{{ time }}</p>
-        <p v-if="canDelete" class="text-xs" @click="clickDeleteBtn">삭제</p>
+        <p class="text-xs mr-1">{{ new_time }}</p>
+        <p v-if="canDelete" class="text-xs text-zz-negative" @click="clickDeleteBtn">· 삭제</p>
       </div>
       <p class="text-base mb-1">{{ content }}</p>
     </li>
@@ -18,14 +18,16 @@
 import { reactive, toRefs } from '@vue/reactivity';
 import { deleteNestedComment } from '@/api/titleCompetition.js';
 import { computed } from 'vue-demi';
-
+import { useRouter } from 'vue-router';
+import Swal from 'sweetalert2';
 export default {
   name: 'NestedCommentListItem',
   props: {
     nested_comment: Object,
+    index: Number,
   },
-  setup(props) {
-    // TODO: time ~~전으로 출력하기
+  emits: ['popNestedComment'],
+  setup(props, ctx) {
     const nested_comment_data = reactive({
       profile_image: 'profile.jpg',
       username: props.nested_comment.username,
@@ -36,29 +38,74 @@ export default {
       // modified: false,
     });
 
-    // TODO: 나중에 로그인 기능 완성되면 username 수정하기
+    const current_userid = window.localStorage.getItem('current_userid');
+    const router = useRouter();
+
+    const goToProfile = () => {
+      router.push(`/profile/${nested_comment_data.username}`);
+    };
+
+    // 삭제 버튼
     const canDelete = computed(() => {
-      console.log(nested_comment_data.username);
-      return (nested_comment_data.username = 'c109');
+      return (nested_comment_data.username = current_userid);
     });
 
+    // 시간표시: ~ 전
+    const new_time = timeForToday(new Date(nested_comment_data.time));
+    function timeForToday(value) {
+      const today = new Date();
+      const timeValue = new Date(value);
+
+      const betweenTime = Math.floor((today.getTime() - timeValue.getTime()) / 1000 / 60);
+      if (betweenTime < 1) return '방금 전';
+      if (betweenTime < 60) {
+        return `${betweenTime}분 전`;
+      }
+
+      const betweenTimeHour = Math.floor(betweenTime / 60);
+      if (betweenTimeHour < 24) {
+        return `${betweenTimeHour}시간 전`;
+      }
+
+      const betweenTimeDay = Math.floor(betweenTime / 60 / 24);
+      if (betweenTimeDay < 365) {
+        return `${betweenTimeDay}일 전`;
+      }
+
+      return `${Math.floor(betweenTimeDay / 365)}년 전`;
+    }
+
     const clickDeleteBtn = () => {
-      deleteNestedComment(
-        nested_comment_data.nested_comment_id,
-        (({ data }) => {
-          console.log(data);
-          //TODO: 데이터 삭제 후 부모의 nested comment 배열에서 삭제하기
-        },
-        (error) => {
-          console.log(error);
-        }),
-      );
+      Swal.fire({
+        icon: 'warning',
+        text: '삭제하시겠습니까?',
+        showCancelButton: true, // cancel버튼 보이기. 기본은 원래 없음
+        confirmButtonColor: '#3085d6', // confrim 버튼 색깔 지정
+        cancelButtonColor: '#d33', // cancel 버튼 색깔 지정
+        confirmButtonText: '승인', // confirm 버튼 텍스트 지정
+        cancelButtonText: '취소', // cancel 버튼 텍스트 지정
+      }).then((result) => {
+        if (result.isConfirmed) {
+          ctx.emit('popNestedComment', props.index);
+          deleteNestedComment(
+            nested_comment_data.nested_comment_id,
+            (({ data }) => {
+              console.log(data);
+            },
+            (error) => {
+              console.log(error);
+            }),
+          );
+        }
+      });
     };
 
     return {
       ...toRefs(nested_comment_data),
       canDelete,
       clickDeleteBtn,
+      new_time,
+      goToProfile,
     };
   },
 };
